% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chelsa-package.R
\docType{package}
\name{Chelsa-package}
\alias{Chelsa}
\alias{Chelsa-package}
\title{Chelsa: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcpri3/Chelsa}
  \item Report bugs at \url{https://github.com/mcpri3/Chelsa/issues}
}

}
\author{
\strong{Maintainer}: Marie-Caroline Prima \email{marie-caroline.prima@univ-grenoble-alpes.fr} (\href{https://orcid.org/{{orcid}}}{ORCID}) [copyright holder]

}
\keyword{internal}
